Move status chip
